+++ Assets/Scripts/Shooting/BulletScripts/Bullet.cs
@@ -13,7 +13,7 @@ public class Bullet : MonoBehaviour
 
 
     public float bulletSpeed = 10f;
-    public int damage = 10;
+    public float damage = 10f;
 
     public RaycastHit2D hitTheWall(Rigidbody2D rb2d, BoxCollider2D collider)
     {
